requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
praw==7.2.0
prawcore==2.0.0
python-dotenv==0.17.1
requests==2.25.1
update-checker==0.18.0
urllib3==1.26.5
websocket-client==1.0.1
